requirements.txt
Django==2.0
certifi==2017.11.5
chardet==3.0.4
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.6
pytz==2017.3
requests==2.18.4
urllib3==1.22
